requirements.txt
aiogram==2.18
apscheduler==3.9.1
python-dotenv
